requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
black==19.10b0
cachetools==4.1.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
distlib==0.3.1
Django==3.0.7
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-rest-knox==4.1.0
django-six==1.0.4
djangorestframework==3.11.0
filelock==3.0.12
Flask==1.1.2
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
pathspec==0.8.0
pipenv==2020.6.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pytz==2020.1
regex==2020.6.8
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
screen==1.0.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.0.25
virtualenv-clone==0.5.4
Werkzeug==1.0.1
xlwt==1.3.0
